:zap:feat: add update user exception and check password when deleting user

+++ src/main/java/com/sparta/newsfeed/user/service/UserService.java
@@ -3,10 +3,7 @@ package com.sparta.newsfeed.user.service;
 import com.sparta.newsfeed.auth.service.AuthService;
 import com.sparta.newsfeed.config.PasswordEncoder;
 import com.sparta.newsfeed.newsfeed.repository.NewsfeedRepository;
-import com.sparta.newsfeed.user.dto.SignUpRequestDto;
-import com.sparta.newsfeed.user.dto.SignUpResponseDto;
-import com.sparta.newsfeed.user.dto.UserRequestDto;
-import com.sparta.newsfeed.user.dto.UserResponseDto;
+import com.sparta.newsfeed.user.dto.*;
 import com.sparta.newsfeed.user.entity.User;
 import com.sparta.newsfeed.user.repository.UserRepository;
 import jakarta.servlet.http.HttpSession;
@@ -48,8 +45,14 @@ public class UserService {
     }
 
     public UserResponseDto updateUser(Long id, UserRequestDto requestDto, HttpSession session){
+
         //TODO: 에러 처리
         User foundUser = userRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Does not exist id = " + id));
+
+        if (!foundUser.getEmail().equals(requestDto.getEmail()) && userRepository.existsUserByEmail(requestDto.getEmail())) {
+            // 해당 이메일의 사용자가 탈퇴했는지 여부는 알 수 없음
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "가입이 불가능한 이메일입니다.");
+        }
 
         String sessionEmail = authService.getSessionEmail(session);
 
@@ -75,11 +78,16 @@ public class UserService {
         return new UserResponseDto(id, requestDto.getEmail(), requestDto.getName());
     }
 
+    public void deleteUser(Long id, UserDeleteRequestDto requestDto , HttpSession session){
+        User foundUser = userRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Does not exist id = " + id));
 
-    public void deleteUser(Long id, HttpSession session){
+        //TODO:겹치는 코드 , 메서드 추출 ?
+        if (!passwordEncoder.matches(requestDto.getPassword(), foundUser.getPassword())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Password is wrong.");
+        }
+
         newsfeedRepository.deleteNewsfeedsByUserId(id);
         //TODO: 에러 처리
-        User foundUser = userRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Does not exist id = " + id));
         foundUser.updateSoftDelete();
 
         authService.logout(session);
